+++ src/main/java/com/github/merchantpug/apugli/registry/ApugliPowerFactories.java
@@ -14,6 +14,7 @@ public class ApugliPowerFactories {
         register(ActionOnDurabilityChange.getFactory());
         register(ActionOnEquipPower.getFactory());
         register(ActionOnTameHitPower.getFactory());
+        register(ActionOnTargetDeathPower.getFactory());
         register(ActionWhenTameHitPower.getFactory());
         register(Power.createSimpleFactory(AerialAffinityPower::new, Apugli.identifier("aerial_affinity")));
         register(AllowAnvilEnchantPower.getFactory());